core: cells: Split AtomDecomposition implementation from global state mutation.
