+++ src/core/CMakeLists.txt
@@ -42,7 +42,7 @@ set(EspressoCore_SRC
     virtual_sites.cpp
     exclusions.cpp
     CellStructure.cpp
-    PartCfg.cpp)
+    PartCfg.cpp AtomDecomposition.hpp)
 
 if(CUDA)
   set(EspressoCuda_SRC
